--- app/javascript/custom/reiatu_alert.js
@@ -1,9 +1,0 @@
-'use strict';
-
-	const buttons = document.querySelectorAll('.button');
-
-	buttons.forEach(button => {
-		button.addEventListener('click', () => {
-			alert("ログイン後にいいね機能が使えます");          
-		});    
-	});
